+++ app/board.py
@@ -1,58 +1,5 @@
 
-
-
-#COLS_MAP = {"A": 0, "B": 1 , "C": 2}
-
-#def row_col(grid_notation):
-#    """Params cell notation like "A1",  where and cols are A, B, C and rows are 1, 2, 3
-#
-#           A | B | C
-#        1  _ | _ | _
-#        2  _ | _ | _
-#        3  _ | _ | _
-#
-#    """
-#    grid_notation = grid_notation.upper() #> "A1"
-#    col, row = list(grid_notation)
-#    col_idx = COLS_MAP[col]
-#    row_idx = int(row) - 1
-#    return row_idx, col_idx
-
-
-#PLAYER_NONE = " "
-#PLAYER_X = "X"
-#PLAYER_O = "O"
-
-#SQUARE_NAMES = [
-#    "A1", "B1", "C1",
-#    "A2", "B2", "C2",
-#    "A3", "B3", "C3",
-#]
-
-class Square:
-    def __init__(self, name):
-        self.name = name
-        self.player_name = None
-
-    def __repr__(self):
-        #return f"<Square {self.name} : '{self.player_label}'>"
-        return f"<Square {self.name}>"
-
-    @property
-    def player_label(self):
-        return self.player_name or " "
-
-    #def row_name(self):
-    #    return "A"
-    #
-    #def col_name(self):
-    #    return "1"
-    #
-    #def row_idx(self):
-    #    return 0
-    #
-    #def col_idx(self):
-    #    return 0
+from app.square import Square
 
 # if a single player has any of these combination of squares, they win:
 WINNING_COMBINATIONS = [
@@ -67,15 +14,19 @@ WINNING_COMBINATIONS = [
 ]
 
 class Board:
-
     def __init__(self):
-        #self.squares = [Square(square_name) for square_name in SQUARE_NAMES]
-
         self.squares = [
             Square("A1"), Square("B1"), Square("C1"),
             Square("A2"), Square("B2"), Square("C2"),
             Square("A3"), Square("B3"), Square("C3"),
         ]
+        # consider this instead, for computational efficiency:
+        # self.squares = {
+        #     "A1": Square("A1"),
+        #     "B1": Square("B1")
+        #     # ...
+        # }
+        # then update self.get_square to use a dictionary-based approach
 
     def __repr__(self):
         return f"""
@@ -89,14 +40,9 @@ class Board:
 
         """
 
-
     @property
     def selectable_squares(self):
         return [square for square in self.squares if not square.player_name]
-
-    #def parse_notation(self, grid_notation):
-    #    return grid_notation.upper()[0:2]
-
 
     def get_square(self, square_name):
         #return self.squares[square_name]
@@ -121,23 +67,6 @@ class Board:
 
     def get_squares(self, square_names):
         return [square for square in self.squares if square.name in square_names]
-
-
-    #@property
-    #def is_solved(self):
-    #    # TRUE if the same player has all squares in any one of the winning combos
-#
-    #    for square_names in WINNING_COMBINATIONS:
-    #        squares = self.get_squares(square_names)
-    #        player_names = [square.player_name for square in squares] #> ['X', None, None]
-    #        if len(set(player_names)) == 1:
-    #            return True
-    #    return False
-
-
-    #@property
-    #def is_solved(self):
-    #    return True if self.winning_player_name else False
 
     @property
     def winning_player_name(self):
